// RehniMarket-frontend/src/types/company.ts
export interface DashboardHomeResponse {
  logo: string;
  banner: string;
  nameCompany: string;
  addressCompany: string;
  emailCompany: string;
  CompanyCertificate: boolean;
  memberAT: string;
  role: string;
  sales: number;
  stars: number;
  reviews: number;
}

export interface CompanyProfileResponse {
  nameCompany: string;
  emailCompany: string;
  addressCompany: string;
  tellCompany: string;
  memberAT: string;

  averageRating: number;
  totalReviews: number;
  completeSales: number;
  sellerLevel: string;
  logo?: string;
  banner?: string;
}

export interface UpdateProfileRequest{
  nameCompany: string;
  emailCompany: string;
  addressCompany: string;
  tellCompany: string;
}

export interface CompanyMediaUpload {
  photo_profile?: File;
  banner_profile?: File;
}


export interface ProductSpecification {
  specificationTemplateId: string;
  value: string;
}

export interface ProductImage {
  file: File;
  preview: string;
  isMain: boolean;
}

export interface CreateProductRequest {
  nameProduct: string;
  catalogId: string;
  priceProduct: number | "";
  stockProduct: number | "";
  descripcionProduct: string;
  technicalSpecProduct: ProductSpecification[];
  imagesProduct: ProductImage[];
}

export interface CatalogResponse {
  id: string;
  name: string;
}

export interface SpecificationResponse {
  id: string;
  name: string;
  type: string;
  required: boolean;
}


export interface MyProductResponse {
  id: string;
  name: string;
  description: string;
  category: string;
  price: number;
  stock: number;
  image: string | null;
  is_active: boolean;
}


export interface MyProductsPaginationResponse {
  page: number;
  limit: number;
  total: number;
  total_pages: number;
  products: MyProductResponse[];
}

export interface ChangeProductStatus {
  is_active: boolean,
}